backend/requirements.txt
djangorestframework
PyJWT
PyPDF2